fix(tests): Eliminate real AWS calls from unit tests for fast, reliable runs
All unit tests now complete in under 30s with zero AWS network calls.

Key changes:
- daemon handlers: replace awsManager.ListInstances() with stateManager.LoadState()
  for read operations across rightsizing, idle, project, instance, template, snapshot handlers
- daemon handlers: add testMode guards to AMI, snapshot, cost, marketplace, security,
  volume handlers that inherently require AWS (create/modify/delete)
- rightsizing: fix NaN division bug (0/0 SavingsPercentage when totalDailyCost == 0)
- idle_handlers: remove unused aws import after replacing with state manager
- instance_handlers: skip ?refresh=true AWS refresh in test mode
- template_handlers: skip AMI discovery initialization in test mode
- test_helpers: expose testMode flag and improve test server setup
- cli: fix TestSimplified_AvailableCommands map iteration race - command_resume and
  command_delete both used "stopped-instance"; delete removes it from mock making
  resume fail non-deterministically; changed resume to use "test-instance"
- templates: fix CommunityMetadataStore concurrent access and usage stats tests
- ratelimit: add missing unit tests, fix integration test flakiness
- invitation: fix validation test edge cases
- aws/manager: fix storage error handling test assertions

+++ pkg/ratelimit/ratelimiter.go
@@ -140,9 +140,11 @@ func (tb *TokenBucket) Allow(ctx context.Context) error {
 	tb.updateSuccessRate()
 
 	// Calculate retry after duration
+	// Multiply by float64(time.Second) before converting to time.Duration
+	// to preserve sub-second precision (time.Duration is in nanoseconds)
 	tokensNeeded := 1.0 - tb.tokens
 	secondsPerToken := 60.0 / tb.rate
-	retryAfter := time.Duration(tokensNeeded*secondsPerToken) * time.Second
+	retryAfter := time.Duration(tokensNeeded * secondsPerToken * float64(time.Second))
 
 	return &RateLimitError{
 		Rate:          tb.rate,
